fix globals.cuh

+++ globals.cuh
@@ -4,8 +4,8 @@ enum ALG_TYPE {
     BFS, SSSP, CC, PR
 };
 typedef uint SIZE_TYPE;
-//typedef unsigned long long EDGE_POINTER_TYPE;
-typedef ull EDGE_POINTER_TYPE;
+typedef unsigned long long EDGE_POINTER_TYPE;
+//typedef ull EDGE_POINTER_TYPE;
 
 struct EdgeWithWeight {
     uint toNode;
